FU-217 refactor: entity를 dto로 변환해주는 로직을 클래스 분리를 통해 재사용성을 높임
